+++ view/tws_administration_view/lib/view/articles/trucks/options/trucks_article_table_adapter.dart
@@ -5,9 +5,21 @@ final SessionStorage _sessionStorage = SessionStorage.i;
 final class _TableAdapter implements TWSArticleTableAdapter<Truck> {
   const _TableAdapter();
 
+  String getPlates(Truck item){
+    String plates = '---';
+    if (item.plates.isNotEmpty) {
+      plates = '';
+      for (int cont = 0; cont < item.plates.length; cont++) {
+        plates += item.plates[cont].identifier;
+        if (item.plates.length > cont) plates += '\n';
+      }
+    }
+    return plates;
+  }
+
   @override
   Future<SetViewOut<Truck>> consume(int page, int range, List<SetViewOrderOptions> orderings) async {
-    final SetViewOptions<Truck> options = SetViewOptions<Truck>(false, page, range, null, orderings, <SetViewFilterNodeInterface<Truck>>[]);
+    final SetViewOptions<Truck> options = SetViewOptions<Truck>(false, range, page, null, orderings, <SetViewFilterNodeInterface<Truck>>[]);
     String auth = _sessionStorage.session!.token;
     MainResolver<SetViewOut<Truck>> resolver = await Sources.administration.trucks.view(options, auth);
 
@@ -22,18 +34,72 @@ final class _TableAdapter implements TWSArticleTableAdapter<Truck> {
   
   @override
   TWSArticleTableEditor? composeEditor(Truck set, Function closeReinvoke, BuildContext context) {
-    // TODO: implement composeEditor
-    throw UnimplementedError();
+    return TWSArticleTableEditor(
+      form: const SizedBox.expand(), 
+      onCancel: () {}
+    );
   }
 
   @override
   Widget composeViewer(Truck set, BuildContext context) {
-    // TODO: implement composeViewer
-    throw UnimplementedError();
+    return SizedBox.expand(
+      child: CSMSpacingColumn(
+        spacing: 12,
+        crossAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          TWSPropertyViewer(
+            label: 'Economic',
+            value: set.truckCommonNavigation?.economic ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'VIN',
+            value: set.vin,
+          ),
+          TWSPropertyViewer(
+            label: 'Carrier',
+            value: set.carrierNavigation?.name ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'Manufacturer',
+            value: set.vehiculeModelNavigation?.manufacturerNavigation?.name ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'Motor',
+            value: set.motor ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'SCT number',
+            value: set.sctNavigation?.number ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'USDOT number',
+            value: set.carrierNavigation?.usdotNavigation?.scac ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'Trimestral Maintenance',
+            value: set.maintenanceNavigation?.trimestral.dateOnlyString ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'Anual Maintenance',
+            value: set.maintenanceNavigation?.anual.dateOnlyString ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'Situation',
+            value: set.truckCommonNavigation?.situationNavigation?.name ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'Location',
+            value: set.truckCommonNavigation?.locationNavigation?.name ?? '---',
+          ),
+          TWSPropertyViewer(
+            label: 'Plates',
+            value: getPlates(set),
+          ),
+        ],
+      ),
+    );
   }
 
   @override
-  void onRemoveRequest(Truck set, BuildContext context) {
-    // TODO: implement onRemoveRequest
-  }
+  void onRemoveRequest(Truck set, BuildContext context) { }
 }