+++ view/tws_administration_view/lib/view/articles/trucks/trucks_article.dart
@@ -1,5 +1,7 @@
 import 'package:csm_view/csm_view.dart';
 import 'package:flutter/material.dart';
+import 'package:http/http.dart';
+import 'package:tws_administration_view/core/extension/datetime.dart';
 import 'package:tws_administration_view/core/router/twsa_routes.dart';
 import 'package:tws_administration_view/data/services/sources.dart';
 import 'package:tws_administration_view/data/storages/session_storage.dart';
@@ -10,6 +12,7 @@ import 'package:tws_administration_view/view/widgets/tws_article_table/tws_artic
 import 'package:tws_administration_view/view/widgets/tws_article_table/tws_article_table_adapter.dart';
 import 'package:tws_administration_view/view/widgets/tws_article_table/tws_article_table_agent.dart';
 import 'package:tws_administration_view/view/widgets/tws_article_table/tws_article_table_field_options.dart';
+import 'package:tws_administration_view/view/widgets/tws_property_viewer.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 part 'options/trucks_article_table_adapter.dart';
 
@@ -19,6 +22,7 @@ class TrucksArticle extends CSMPageBase {
 
   @override
   Widget compose(BuildContext ctx, Size window) {
+    _TableAdapter adapter = const _TableAdapter();
     return BusinessFrame(
       currentRoute: TWSARoutes.trucksArticle,
       actionsOptions: ActionRibbonOptions(
@@ -27,25 +31,39 @@ class TrucksArticle extends CSMPageBase {
         ),
       ),
       article: TWSArticleTable<Truck>(
-        adapter: const _TableAdapter(),
+        editable:  false,
+        removable: false,
+        adapter: adapter,
         fields: <TWSArticleTableFieldOptions<Truck>>[
+           TWSArticleTableFieldOptions<Truck>(
+            'Economic',
+            (Truck item, int index, BuildContext ctx) => item.truckCommonNavigation?.economic ?? '---',
+          ),
           TWSArticleTableFieldOptions<Truck>(
             'VIN',
             (Truck item, int index, BuildContext ctx) => item.vin,
           ),
-          // TWSArticleTableFieldOptions<Truck>(
-          //   'Manufacturer',
-          //   (Truck item, int index, BuildContext ctx) => item.manufacturerNavigation?.brand ?? '---',
-          // ),
-          // TWSArticleTableFieldOptions<Truck>(
-          //   'Motor',
-          //   (Truck item, int index, BuildContext ctx) => item.motor,
-          //   true,
-          // ),
+          TWSArticleTableFieldOptions<Truck>(
+            'Carrier',
+            (Truck item, int index, BuildContext ctx) => item.carrierNavigation?.name ?? '---',
+          ),
+          TWSArticleTableFieldOptions<Truck>(
+            'Manufacturer',
+            (Truck item, int index, BuildContext ctx) => item.vehiculeModelNavigation?.manufacturerNavigation?.name ?? '---',
+          ),
+          TWSArticleTableFieldOptions<Truck>(
+            'Motor',
+            (Truck item, int index, BuildContext ctx) => item.motor ?? '---',
+            true,
+          ),
           TWSArticleTableFieldOptions<Truck>(
             'SCT Number',
-            (Truck item, int index, BuildContext ctx) =>  '---',
+            (Truck item, int index, BuildContext ctx) =>  item.sctNavigation?.number ?? '---',
             true,
+          ),
+          TWSArticleTableFieldOptions<Truck>(
+            'USDOT number',
+            (Truck item, int index, BuildContext ctx) => item.carrierNavigation?.usdotNavigation?.scac ?? '---',
           ),
           TWSArticleTableFieldOptions<Truck>(
             'Trimestral Maintenance',
@@ -58,30 +76,26 @@ class TrucksArticle extends CSMPageBase {
             true,
           ),
           TWSArticleTableFieldOptions<Truck>(
-            'Situation',
-            (Truck item, int index, BuildContext ctx) => '---',
-            true,
-          ),
-          TWSArticleTableFieldOptions<Truck>(
             'Insurance',
             (Truck item, int index, BuildContext ctx) => item.insuranceNavigation?.policy ?? '---',
             true,
           ),
           TWSArticleTableFieldOptions<Truck>(
             'Plates',
-            (Truck item, int index, BuildContext ctx) {
-              String plates = '---';
-              if (item.plates.isNotEmpty) {
-                plates = '';
-                for (int cont = 0; cont < item.plates.length; cont++) {
-                  plates += item.plates[cont].identifier;
-                  if (item.plates.length > cont) plates += '\n';
-                }
-              }
-              return plates;
-            },
+            (Truck item, int index, BuildContext ctx) => adapter.getPlates(item),
             true,
           ),
+          TWSArticleTableFieldOptions<Truck>(
+            'Situation',
+            (Truck item, int index, BuildContext ctx) => item.truckCommonNavigation?.situationNavigation?.name ?? '---',
+            true,
+          ),
+          TWSArticleTableFieldOptions<Truck>(
+            'Location',
+            (Truck item, int index, BuildContext ctx) => item.truckCommonNavigation?.locationNavigation?.name ?? '---',
+            true,
+          ),
+          
         ],
         page: 1,
         size: 25,