+++ view/tws_administration_view/lib/view/frames/business/business_frame.dart
@@ -30,26 +30,6 @@ class BusinessFrame extends StatelessWidget {
       route: TWSARoutes.trucksArticle,
       title: 'Trucks',
     ),
-    ArticleOptions(
-      icon: (Color? stateColor) {
-        return Icon(
-          Icons.business,
-          color: stateColor,
-        );
-      },
-      route: TWSARoutes.manufacturersArticle,
-      title: 'Manufacturers',
-    ),
-    ArticleOptions(
-      icon: (Color? stateColor) {
-        return Icon(
-          Icons.business,
-          color: stateColor,
-        );
-      },
-      route: TWSARoutes.situationsArticle,
-      title: 'Situations',
-    ),
   ];
 
   @override