+++ view/tws_administration_view/CHANGELOG.md
@@ -11,7 +11,9 @@
         * [TWSDatepickerField]
         * [TWSInputText]
     3. Removed unnecesary whispers views:
-        *
+        * Manufacturers
+        * Situations
+    4. [Truck] view table and table viewer implementation
 
 - Dependencies upgrade: